+++ CHANGELOG.md
@@ -53,3 +53,5 @@
 * *[/Beta1.rc11]*
 * *[Beta1.rc12]*
   * [FIX]: Removed chrome spacer
+  * [FEATURE]: Notify parent of player preview modal opening as a state
+    change